+++ Game_2D/Point2D.cpp
@@ -22,11 +22,11 @@ void Point2D::changeY(int y) {
 }
 
 void Point2D::draw(sf::RenderWindow& window) {
-	primitive.drawPoint(window,x,y);
+	primitive.drawPoint(window, x, y);
 }
 
 void Point2D::draw(sf::RenderWindow& window, sf::Color color) {
-	primitive.drawPoint(window, x, y,color);
+	primitive.drawPoint(window, x, y, color);
 }
 
 void Point2D::translate(sf::RenderWindow& window, int x, int y) {
@@ -41,7 +41,7 @@ void Point2D::rotate(sf::RenderWindow& window, float angle) {
 	this->y = (int)y2;
 }
 
-void Point2D::scale(sf::RenderWindow& window, int x, int y) {
+void Point2D::scale(sf::RenderWindow& window, float x, float y) {
 	float x2 = this->x * x;
 	float y2 = this->y * y;
